+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -307,6 +307,12 @@ New checks
   Detects issues in accordance with MISRA C rule 17.1:
   "`The features of <stdarg.h> shall not be used`".
 
+- New :doc:`misrac-rule-17p8
+  <clang-tidy/checks/misrac/rule-17p8>` check.
+
+  Detects issues in accordance with MISRA C rule 17.8:
+  "`A function parameter should not be modified`".
+
 - New :doc:`misrac-rule-18p5
   <clang-tidy/checks/misrac/rule-18p5>` check.